requirements.txt
# This file was autogenerated by uv via the following command:
# Original dependencies
docling==2.57.0
langchain==1.0.2
langchain-text-splitters==1.0.0
openai==2.6.0
pdfplumber==0.11.7
pinecone==7.3.0
PyPDF2==3.0.1
plotly==6.3.1
python-dotenv==1.1.1
tiktoken==0.12.0
tomli==2.3.0

# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Streamlit and UI dependencies
streamlit==1.28.1
streamlit-chat==0.1.1

# RAG system dependencies
sentence-transformers==2.2.2
wikipedia-api==0.6.0
rank-bm25==0.2.2
sse-starlette==1.6.5

# Additional utilities
numpy==1.24.3
pandas==2.0.3
asyncio-throttle==1.0.2
